Move api to component and make arguments optional
