+++ nni/retiarii/nn/pytorch/__init__.py
@@ -1,2 +1,3 @@
 from .api import *
+from .component import *
 from .nn import *
