+++ nni/retiarii/nn/pytorch/api.py
@@ -4,17 +4,16 @@
 import copy
 import warnings
 from collections import OrderedDict
-from typing import Any, Callable, List, Union, Dict, Tuple
+from typing import Any, List, Union, Dict
 
 import torch
 import torch.nn as nn
 
 from ...serializer import Translatable, basic_unit
 from ...utils import uid
-from .nn import ModuleList
 
 
-__all__ = ['LayerChoice', 'InputChoice', 'ValueChoice', 'Placeholder', 'ChosenInputs', 'Repeat', 'Cell']
+__all__ = ['LayerChoice', 'InputChoice', 'ValueChoice', 'Placeholder', 'ChosenInputs', ]
 
 
 class LayerChoice(nn.Module):
@@ -311,126 +310,6 @@ class ValueChoice(Translatable, nn.Module):
         for candidate in self.candidates:
             access.access(candidate)
         return access
-
-
-class Repeat(nn.Module):
-    """
-    Repeat a block by a variable number of times.
-
-    Parameters
-    ----------
-    blocks : function, list of function, module or list of module
-        The block to be repeated. If not a list, it will be replicated into a list.
-        If a list, it should be of length ``max_depth``, the modules will be instantiated in order and a prefix will be taken.
-        If a function, it will be called to instantiate a module. Otherwise the module will be deep-copied.
-    depth : int or tuple of int
-        If one number, the block will be repeated by a fixed number of times. If a tuple, it should be (min, max),
-        meaning that the block will be repeated at least `min` times and at most `max` times.
-    """
-    def __init__(self,
-                 blocks: Union[Callable[[], nn.Module], List[Callable[[], nn.Module]], nn.Module, List[nn.Module]],
-                 depth: Union[int, Tuple[int, int]], label=None):
-        super().__init__()
-        self._label = label if label is not None else f'repeat_{uid()}'
-        self.min_depth = depth if isinstance(depth, int) else depth[0]
-        self.max_depth = depth if isinstance(depth, int) else depth[1]
-        assert self.max_depth >= self.min_depth > 0
-        if not isinstance(blocks, list):
-            if isinstance(blocks, nn.Module):
-                blocks = [blocks] + [copy.deepcopy(blocks) for _ in range(self.max_depth - 1)]
-            else:
-                blocks = [blocks for _ in range(self.max_depth)]
-        assert len(blocks) > 0
-        if not isinstance(blocks[0], nn.Module):
-            blocks = [b() for b in blocks]
-        assert self.max_depth <= len(blocks)
-        self.blocks = nn.ModuleList(blocks)
-
-    @property
-    def label(self):
-        return self._label
-
-    def forward(self, x):
-        for block in self.blocks:
-            x = block(x)
-        return x
-
-
-class Cell(nn.Module):
-    """
-    Cell structure [1]_ [2]_ that is popularly used in NAS literature.
-
-    A cell consists of multiple "nodes". Each node is a sum of multiple operators. Each operator is chosen from
-    ``op_candidates``, and takes one input from previous nodes and predecessors. Predecessor means the input of cell.
-    The output of cell is the concatenation of some of the nodes in the cell (currently all the nodes).
-
-    Parameters
-    ----------
-    op_candidates : function or list of module
-        A list of modules to choose from, or a function that returns a list of modules.
-    num_nodes : int
-        Number of nodes in the cell.
-    num_ops_per_node: int
-        Number of operators in each node. The output of each node is the sum of all operators in the node.
-    num_predecessors : int
-        Number of inputs of the cell. The input to forward should be a list of tensors.
-    merge_op : str
-        Currently only ``all`` is supported, which has slight difference with that described in reference.
-    label : str
-        Identifier of the cell. Cell sharing the same label will semantically share the same choice.
-
-    References
-    ----------
-    .. [1] Barret Zoph, Quoc V. Le, "Neural Architecture Search with Reinforcement Learning". https://arxiv.org/abs/1611.01578
-    .. [2] Barret Zoph, Vijay Vasudevan, Jonathon Shlens, Quoc V. Le,
-        "Learning Transferable Architectures for Scalable Image Recognition". https://arxiv.org/abs/1707.07012
-    """
-
-    # TODO:
-    # Support loose end concat (shape inference on the following cells)
-    # How to dynamically create convolution with stride as the first node
-
-    def __init__(self,
-                 op_candidates: Union[Callable, List[nn.Module]],
-                 num_nodes: int,
-                 num_ops_per_node: int,
-                 num_predecessors: int,
-                 merge_op: str,
-                 label: str = None):
-        super().__init__()
-        self._label = f'cell_{uid()}' if label is None else label
-        self.ops = ModuleList()
-        self.inputs = ModuleList()
-        self.num_nodes = num_nodes
-        self.num_ops_per_node = num_ops_per_node
-        self.num_predecessors = num_predecessors
-        for i in range(num_nodes):
-            self.ops.append(ModuleList())
-            self.inputs.append(ModuleList())
-            for k in range(num_ops_per_node):
-                if isinstance(op_candidates, list):
-                    assert len(op_candidates) > 0 and isinstance(op_candidates[0], nn.Module)
-                    ops = copy.deepcopy(op_candidates)
-                else:
-                    ops = op_candidates()
-                self.ops[-1].append(LayerChoice(ops, label=f'{self.label}__op_{i}_{k}'))
-                self.inputs[-1].append(InputChoice(i + num_predecessors, 1, label=f'{self.label}/input_{i}_{k}'))
-        assert merge_op in ['all']  # TODO: loose_end
-        self.merge_op = merge_op
-
-    @property
-    def label(self):
-        return self._label
-
-    def forward(self, x: List[torch.Tensor]):
-        states = x
-        for ops, inps in zip(self.ops, self.inputs):
-            current_state = []
-            for op, inp in zip(ops, inps):
-                current_state.append(op(inp(states)))
-            current_state = torch.sum(torch.stack(current_state), 0)
-            states.append(current_state)
-        return torch.cat(states[self.num_predecessors:], 1)
 
 
 @basic_unit
